+++ source/rust/code/math/algebra/matrix/ntm_matrix_transpose.rs
@@ -42,9 +42,9 @@
 //                                                                               //
 ///////////////////////////////////////////////////////////////////////////////////
 
-pub fn ntm_matrix_transpose(matrix: Vec<Vec<i32>>) -> Vec<Vec<i32>> {
+pub fn ntm_matrix_transpose(matrix: Vec<Vec<f64>>) -> Vec<Vec<f64>> {
     // Transpose a matrix of any size
-    let mut result: Vec<Vec<i32>> = vec![Vec::with_capacity(matrix.len()); matrix[0].len()];
+    let mut result: Vec<Vec<f64>> = vec![Vec::with_capacity(matrix.len()); matrix[0].len()];
 
     for row in matrix {
         for column in 0..row.len() {
@@ -55,11 +55,11 @@ pub fn ntm_matrix_transpose(matrix: Vec<Vec<i32>>) -> Vec<Vec<i32>> {
 }
 
 fn main() {
-    let input0: Vec<Vec<i32>> = vec![vec![1, 0, 1], vec![0, 2, 0], vec![5, 0, 1]];
-    let input1: Vec<Vec<i32>> = vec![vec![3, 4, 2], vec![0, 1, 3], vec![3, 1, 1]];
+    let input0: Vec<Vec<f64>> = vec![vec![1.0, 0.0, 1.0], vec![0.0, 2.0, 0.0], vec![5.0, 0.0, 1.0]];
+    let input1: Vec<Vec<f64>> = vec![vec![3.0, 4.0, 2.0], vec![0.0, 1.0, 3.0], vec![3.0, 1.0, 1.0]];
 
-    let output0: Vec<Vec<i32>> = vec![vec![1, 0, 5], vec![0, 2, 0], vec![1, 0, 1]];
-    let output1: Vec<Vec<i32>> = vec![vec![3, 0, 3], vec![4, 1, 1], vec![2, 3, 1]];
+    let output0: Vec<Vec<f64>> = vec![vec![1.0, 0.0, 5.0], vec![0.0, 2.0, 0.0], vec![1.0, 0.0, 1.0]];
+    let output1: Vec<Vec<f64>> = vec![vec![3.0, 0.0, 3.0], vec![4.0, 1.0, 1.0], vec![2.0, 3.0, 1.0]];
 
     assert_eq!(ntm_matrix_transpose(input0), output0);
     assert_eq!(ntm_matrix_transpose(input1), output1);
