+++ source/rust/code/math/algebra/matrix/ntm_matrix_vector_product.rs
@@ -42,9 +42,9 @@
 //                                                                               //
 ///////////////////////////////////////////////////////////////////////////////////
 
-pub fn ntm_matrix_vector_product(multiplier: Vec<i32>, multiplicand: Vec<i32>) -> Vec<Vec<i32>> {
+pub fn ntm_matrix_vector_product(multiplier: Vec<f64>, multiplicand: Vec<f64>) -> Vec<Vec<f64>> {
     // Multiply two matching matrices.
-    let mut result: Vec<Vec<i32>> = vec![];
+    let mut result: Vec<Vec<f64>> = vec![];
 
     // The multiplier needs to have the same amount of columns as the multiplicand has rows.
     let mut temporal;
@@ -60,14 +60,14 @@ pub fn ntm_matrix_vector_product(multiplier: Vec<i32>, multiplicand: Vec<i32>) -
 }
 
 fn main() {
-    let input_a: Vec<i32> = vec![1, 2, 3, 4];
-    let input_b: Vec<i32> = vec![1, 3, 2];
+    let input_a: Vec<f64> = vec![1.0, 2.0, 3.0, 4.0];
+    let input_b: Vec<f64> = vec![1.0, 3.0, 2.0];
 
-    let output: Vec<Vec<i32>> = vec![
-        vec![1, 3, 2],
-        vec![2, 6, 4],
-        vec![3, 9, 6],
-        vec![4, 12, 8]
+    let output: Vec<Vec<f64>> = vec![
+        vec![1.0, 3.0, 2.0],
+        vec![2.0, 6.0, 4.0],
+        vec![3.0, 9.0, 6.0],
+        vec![4.0, 12.0, 8.0]
     ];
 
     assert_eq!(ntm_matrix_vector_product(input_a, input_b), output);
