+++ source/rust/code/math/algebra/scalar/ntm_scalar_multiplication.rs
@@ -42,8 +42,8 @@
 //                                                                               //
 ///////////////////////////////////////////////////////////////////////////////////
 
-pub fn ntm_scalar_multiplication(vector: Vec<i32>) -> i32 {
-    let mut result: i32 = 1;
+pub fn ntm_scalar_multiplication(vector: Vec<f64>) -> f64 {
+    let mut result: f64 = 1.0;
 
     for row in 0..vector.len() {
         result *= vector[row];
@@ -52,11 +52,11 @@ pub fn ntm_scalar_multiplication(vector: Vec<i32>) -> i32 {
 }
 
 fn main() {
-    let input0: Vec<i32> = vec![3, 2, 2];
-    let input1: Vec<i32> = vec![1, 0, 0];
+    let input0: Vec<f64> = vec![3.0, 2.0, 2.0];
+    let input1: Vec<f64> = vec![1.0, 0.0, 0.0];
 
-    let output0: i32 = 12;
-    let output1: i32 = 0;
+    let output0: f64 = 12.0;
+    let output1: f64 = 0.0;
 
     assert_eq!(ntm_scalar_multiplication(input0), output0);
     assert_eq!(ntm_scalar_multiplication(input1), output1);
